my-leaderboard/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
PyJWT==2.8.0
sqlalchemy==2.0.25
pydantic==2.5.3
scipy==1.13.0
numpy==1.26.4
python-multipart==0.0.6
requests==2.31.0
pytest==8.3.3
httpx==0.25.2
datasets==2.19.1

# Infrastructure
redis==5.0.1
slowapi==0.1.9
python-json-logger==2.0.7
cachetools==5.3.2

# Database drivers
psycopg2-binary==2.9.9

